refactor: move themeProvider to providers folder
