+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import './global.css';
 import Menu from '@/components/menu/Menu';
 import ReduxProvider from '@/app/providers/ReduxProviders';
-import { ThemeProvider } from '@/components/theme-provider/ThemeProvider';
+import { ThemeProvider } from '@/app/providers/ThemeProvider';
 
 export default function LocaleLayout({
   children,
