+++ src/app/providers/ThemeProvider.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { type ReactNode } from 'react';
-import { ThemeContext } from '../../context/ThemeContext';
-import { useLocalStorage } from '../../hooks/useLocalStorage';
+import { ThemeContext } from '../context/ThemeContext';
+import { useLocalStorage } from '../hooks/useLocalStorage';
 
 export function ThemeProvider({ children }: { children: ReactNode }) {
   const [theme, setTheme] = useLocalStorage('theme', 'light');
